requirements.txt
hydralit==1.0.12
nltk==3.6.5
numpy==1.20.3
pandas==1.3.4
Pillow==9.0.1
plotly==5.6.0
pyrebase==3.0.27
Pyrebase4==4.5.0
pyspellchecker==0.6.3
scikit_learn==1.0.2
seaborn==0.11.2
spellchecker==0.4
st_annotated_text==3.0.0
streamlit==1.6.0
streamlit_authenticator==0.1.0
textblob==0.17.1
validate_email==1.3
